MAINT: Replace old ema_talib example with new one.
